// Admin/src/app/core/models/staff.models.ts
export interface staff {
    id: number;
    fname: string;
    lname: string;
    gender: string;
    username: string;
    email: string;
    password: string;
    clinicId: string;
    roleid: string;
    desgid: string;
    deptId: string;
    dob: string;
    doj: string;
    maritalstatus: string;
    province: string;
    district: string;
    subdistrict: string;
    qualification: string;
    sancno: string;
    hpcsano: string;
    contactno: string;
    profileimg: string;
    address: string;
    staffType: string;
    level: string;
    added_by: string;
    status: string;
}